修复立方体纯黑问题（是因为 shader setVec3() 的参数设置为 int 类型导致数据截断），让光源颜色随时间变化并反映在立方体上。

+++ OpenGL/OpenGL/shader.cpp
@@ -102,7 +102,7 @@ void Shader::setMat4(const std::string& name, const glm::mat4& value) const
     glUniformMatrix4fv(glGetUniformLocation(shader_program, name.c_str()), 1, GL_FALSE, glm::value_ptr(value));
 }
 
-void Shader::setVec3(const std::string& name, int x, int y, int z) const
+void Shader::setVec3(const std::string& name, float x, float y, float z) const
 {
     glUniform3f(glGetUniformLocation(shader_program, name.c_str()), x, y, z);
 }
